app/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
alembic==1.10.4
anyio==4.3.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
bcrypt==4.3.0
certifi==2024.2.2
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.0
cryptography==44.0.3
cycler==0.12.1
decorator==5.1.1
dnspython==2.7.0
ecdsa==0.18.0
email_validator==2.2.0
et-xmlfile==1.1.0
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.89.1
fonttools==4.49.0
frozenlist==1.4.1
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.6
iniconfig==2.1.0
ipython==8.23.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.6
joblib==1.3.2
kiwisolver==1.4.5
Mako==1.3.2
MarkupSafe==2.1.5
matplotlib==3.6.2
matplotlib-inline==0.1.7
multidict==6.0.5
mysql-connector-python==8.0.29
numpy==1.26.4
openai==0.27.0
openpyxl==3.1.2
orjson==3.10.18
packaging==23.2
pandas==2.0.3
pandas-stubs==2.2.0.240218
parso==0.8.4
passlib==1.7.4
pexpect==4.9.0
pillow==10.2.0
plotly==5.19.0
pluggy==1.5.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.1
pycparser==2.22
pydantic==1.10.14
Pygments==2.17.2
PyJWT==2.7.0
pyparsing==3.1.1
pytest==8.3.5
python-dateutil==2.8.2
python-dotenv==0.19.2
python-jose==3.3.0
python-multipart==0.0.20
pytz==2024.1
PyYAML==6.0.2
requests==2.31.0
rsa==4.9
scikit-learn==1.4.1.post1
scipy==1.12.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.0
stack-data==0.6.3
starlette==0.22.0
tenacity==8.2.3
threadpoolctl==3.3.0
tomli==2.2.1
tqdm==4.66.2
traitlets==5.14.3
types-pytz==2024.1.0.20240203
typing_extensions==4.10.0
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.1
uvicorn==0.20.0
uvloop==0.21.0
watchfiles==1.0.5
wcwidth==0.2.13
websockets==15.0.1
yarl==1.9.4
